Refactor: Refactor resume repository
