--- server/src/main/java/project/manager/server/repository/ResumeEduRepository.java
@@ -1,8 +1,0 @@
-package project.manager.server.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import project.manager.server.domain.ResumeEdu;
-
-public interface ResumeEduRepository extends JpaRepository<ResumeEdu, Long> {
-
-}
