+++ server/src/main/java/project/manager/server/repository/resume/ProjectRepository.java
@@ -1,0 +1,12 @@
+package project.manager.server.repository.resume;
+
+import java.util.List;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import project.manager.server.domain.resume.Project;
+
+@Repository
+public interface ProjectRepository extends JpaRepository<Project, Long> {
+    Optional<List<Project>> findByResumeId(Long resumeId);
+}
