+++ server/src/main/java/project/manager/server/repository/resume/ResumeRepository.java
@@ -1,15 +1,14 @@
-package project.manager.server.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import project.manager.server.domain.Resume;
+package project.manager.server.repository.resume;
 
 import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import project.manager.server.domain.resume.Resume;
 
 @Repository
 public interface ResumeRepository extends JpaRepository<Resume, Long> {
 
     Optional<Resume> findByUserId(Long id);
+
     boolean existsByUserId(Long id);
 }
-
